pyproject.toml
[project]
name = "django-auth"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "alex-candido", email = "alex.candido.tec@gmail.com"},
]
dependencies = [
    "django>=5.2", 
    "djangorestframework>=3.16.0", 
    "django-extensions>=4.1", 
    "markdown>=3.8", 
    "django-filter>=25.1", 
    "dataclasses>=0.6", 
    "djangorestframework-simplejwt>=5.3.1", 
    "djangorestframework-gis>=1.0",
    "django-cors-headers>=4.7.0", 
    "django-sendgrid-v5>=1.2.3",
    "python-dotenv>=1.0.1", 
    "psycopg2-binary>=2.9.10", 
    "pydantic-settings>=2.9.1", 
    "dj-database-url>=2.3.0", 
    "faker>=37.1.0", 
    "django-rest-passwordreset>=1.5.0", 
    "dj-rest-auth>=7.0.1", 
    "django-allauth>=65.3.0", 
    "requests>=2.32.3", 
    "pygraphviz>=1.14", 
    "pydot>=4.0.0", 
    "django-schema-viewer>=0.5.3", 
    "django-dbml>=0.10.1", 
    "dependency-injector>=4.46.0"
]

requires-python = ">=3.11.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.scripts]
manage="python ./src/manage.py"
server = "pdm run manage runserver 0.0.0.0:8000"
